requirements.txt
torch==1.6.0
torch-cluster==1.5.9
torch-geometric==1.7.0
torch-scatter==2.0.6
torch-sparse==0.6.9
torch-spline-conv==1.2.1
numpy==1.20.3
scikit-learn==1.0.1
pyro-ppl==1.3.0
pyyaml==6.0
